+++ include/CeLux/conversion/cpu/CPUConverters.hpp
@@ -4,7 +4,7 @@
 
 #include <cpu/CPUConverter.hpp>
 #include <cpu/YUV420PToRGB.hpp>
-#include <cpu/RGBToYUV420P.hpp>
 #include <cpu/YUV420P10ToRGB48.hpp>
-#include <cpu/RGB48ToYUV420P10.hpp>
+#include <cpu/BGRToRGB.hpp>
+#include <cpu/RGBToRGB.hpp>
 #endif // CPU_CONVERTERS_HPP