+++ include/CeLux/conversion/gpu/GPUConverters.hpp
@@ -5,9 +5,7 @@
 #include "IConverter.hpp"
 #include "cuda/BaseConverter.hpp"
 #include "cuda/NV12ToRGB.hpp"
-#include "cuda/RGBToYUV420P.hpp"
 #include "cuda/P010LEToRGB.hpp"
-#include "cuda/RGB48ToP010LE.hpp"
 
 
 #endif