+++ tests/benchmarks/test_benchmarks.py
@@ -1,5 +1,3 @@
-# tests/benchmarks/test_benchmarks.py
-
 import pytest
 import torch
 import json
@@ -18,9 +16,8 @@ logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(
 logger = logging.getLogger(__name__)
 
 # Define paths to sample videos as strings
-SAMPLE_VIDEO_CUDA = "tests/data/BigBuckBunny.mp4"      # 14315 frames
-SAMPLE_VIDEO_CPU = "tests/data/BigBuckBunny.mp4"      # 14315 frames
-OUTPUT_VIDEO = "tests/data/output_benchmark.mp4"
+SAMPLE_VIDEO_CUDA = "tests/data/BigBuckBunny.mp4"  # 14315 frames
+SAMPLE_VIDEO_CPU = "tests/data/BigBuckBunny.mp4"   # 14315 frames
 
 @pytest.fixture(scope="module")
 def video_properties():
@@ -36,8 +33,7 @@ def video_properties():
     
     total_frames = {
         "test_video_reader_cpu_benchmark": props_cpu['total_frames'],
-        "test_video_reader_cuda_benchmark": props_cuda['total_frames'],
-        "test_video_writer_benchmark": props_cpu['total_frames']
+        "test_video_reader_cuda_benchmark": props_cuda['total_frames']
     }
 
     # Write to total_frames.json
@@ -84,35 +80,5 @@ def test_video_reader_cuda_benchmark(benchmark, video_properties):
     
     benchmark(read_video)
     assert True  # Dummy assertion to ensure the test doesn't fail
-
-def test_video_writer_benchmark(benchmark, video_properties):
-    """
-    Benchmark the VideoWriter by writing frames to a new video.
-    """
-    def write_video():
-        logger.info("Starting VideoWriter benchmark.")
-        with celux.VideoWriter(
-            OUTPUT_VIDEO,
-            width=video_properties["cuda"]['width'],
-            height=video_properties["cuda"]['height'],
-            fps=video_properties["cuda"]['fps'],
-            device="cuda"
-        ) as writer:
-            with celux.VideoReader(SAMPLE_VIDEO_CPU, device="cuda") as reader:
-                for frame in reader:
-                    writer.write_frame(frame)
-        logger.info("Completed VideoWriter benchmark.")
-    
-    benchmark(write_video)
-    
-    # Verify the output video
-    with celux.VideoReader(OUTPUT_VIDEO, device="cuda") as output_reader:
-        assert len(output_reader) == video_properties["cuda"]['total_frames']
-    
-    # Clean up the output video after benchmarking
-    if Path(OUTPUT_VIDEO).exists():
-        Path(OUTPUT_VIDEO).unlink()
-        logger.info("Cleaned up output video.")
-
 
 #pytest tests\benchmarks\ --benchmark-only --benchmark-json=benchmark_results.json --html=benchmark_report.html